create tabs and dropdown storybook.
