+++ src/components/buttons/Tabs.tsx
@@ -1,12 +1,12 @@
 import styled, { css } from 'styled-components';
 
-const Tabs = styled.div`
+const Wrapper = styled.div`
   margin-right: 8px;
 `;
 
 const activeStyle = css`
-  background-color: ${({ theme }) => theme.emphasis};
-  border: 1px solid ${({ theme }) => theme.emphasis};
+  background-color: #0969da;
+  border: 1px solid #0969da;
   color: #fff;
 `;
 
@@ -17,15 +17,15 @@ const Tab = styled.button<TabProp>`
   padding: 5px 16px;
   font-weight: 500;
   border-radius: 6px;
-  background-color: ${({ theme }) => theme.default};
+  background-color: #f6f8fa;
   color: #24292f;
-  border: 1px solid ${({ theme }) => theme.border};
+  border: 1px solid #d0d7de;
   height: 32px;
   cursor: pointer;
   :hover {
     background-color: #f3f4f6;
   }
-  background-color: ${({ theme }) => theme.light};
+  background-color: #fff;
   ${({ active }) => active && activeStyle};
   border-radius: ${({ index }) =>
     index === 0 ? '6px 0 0 6px' : '0 6px 6px 0'};
@@ -36,17 +36,17 @@ const Text = styled.span`
   line-height: 20px;
 `;
 
-const ButtonGroup = ({ buttons }) => {
+const Tabs = ({ buttons }) => {
   return (
-    <Tabs>
+    <Wrapper>
       {buttons.map((button, index) => (
         <Tab key={index} active={button.active} index={index}>
           {button.icon}
           <Text>{button.text}</Text>
         </Tab>
       ))}
-    </Tabs>
+    </Wrapper>
   );
 };
 
-export default ButtonGroup;
+export default Tabs;
