+++ src/stories/header.css
@@ -21,10 +21,6 @@ h1 {
   vertical-align: top;
 }
 
-button + button {
-  margin-left: 10px;
-}
-
 .welcome {
   color: #333;
   font-size: 14px;
