+++ src/stories/labelList/Dropdown.stories.tsx
@@ -1,0 +1,28 @@
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+import Dropdown from '../../components/buttons/Dropdown';
+
+// More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
+export default {
+  title: 'Components/Dropdown',
+  component: Dropdown,
+  // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
+  argTypes: {
+    // backgroundColor: { control: 'color' },
+  },
+} as ComponentMeta<typeof Dropdown>;
+
+// More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
+const Template: ComponentStory<typeof Dropdown> = (args) => (
+  <Dropdown {...args} />
+);
+
+export const Default = Template.bind({});
+// More on args: https://storybook.js.org/docs/react/writing-stories/args
+Default.args = {
+  sortList: [
+    'Alphabetically',
+    'Reverse alphabetically',
+    'Most issues',
+    'Fewest issues',
+  ],
+};
